Remove unused credentials/tests, rename vc.rs to verifiable_credential.rs
