+++ crates/credentials/src/verifiable_credential.rs
